Adds Storage Module + Fixes
Added a module that handles the interaction between the Todo List
object and local storage. Also added extra functionality in the Todo
List object for updating Inbox, Today, and This Week tasks based on the
due date. Fixed functions for deleting from arrays in the Todo List
object and the Project object. Fixed functions for finding items in
arrays for the Todo List object and the Project object.